refactor : split model ,component and screen at the main class to separated directory
