+++ sample/src/main/java/com/mushafimad/sampleapp/Util/General.kt
@@ -1,0 +1,14 @@
+package com.mushafimad.sampleapp.Util
+
+object General {
+   fun formatDuration(seconds: Long): String {
+        val hours = seconds / 3600
+        val minutes = (seconds % 3600) / 60
+        val secs = seconds % 60
+        return when {
+            hours > 0 -> "${hours}h ${minutes}m ${secs}s"
+            minutes > 0 -> "${minutes}m ${secs}s"
+            else -> "${secs}s"
+        }
+    }
+}
