+++ sample/src/main/java/com/mushafimad/sampleapp/component/BookmarkItem.kt
@@ -1,0 +1,64 @@
+package com.mushafimad.sampleapp.component
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Delete
+import androidx.compose.material.icons.filled.Favorite
+import androidx.compose.material3.Card
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.mushafimad.core.domain.models.Bookmark
+
+@Composable
+fun BookmarkItem(bookmark: Bookmark, onDelete: () -> Unit) {
+    Card(
+        modifier = Modifier.Companion
+            .fillMaxWidth()
+            .padding(vertical = 4.dp)
+    ) {
+        Row(
+            modifier = Modifier.Companion
+                .fillMaxWidth()
+                .padding(12.dp),
+            verticalAlignment = Alignment.Companion.CenterVertically
+        ) {
+            Icon(
+                imageVector = Icons.Default.Favorite,
+                contentDescription = null,
+                tint = MaterialTheme.colorScheme.primary
+            )
+            Spacer(modifier = Modifier.Companion.width(12.dp))
+            Column(modifier = Modifier.Companion.weight(1f)) {
+                Text(
+                    text = "Surah ${bookmark.chapterNumber}, Ayah ${bookmark.verseNumber}",
+                    style = MaterialTheme.typography.titleMedium
+                )
+                if (bookmark.note.isNotEmpty()) {
+                    Text(
+                        text = bookmark.note,
+                        style = MaterialTheme.typography.bodySmall,
+                        color = MaterialTheme.colorScheme.onSurfaceVariant
+                    )
+                }
+            }
+            IconButton(onClick = onDelete) {
+                Icon(
+                    imageVector = Icons.Default.Delete,
+                    contentDescription = "Delete",
+                    tint = MaterialTheme.colorScheme.error
+                )
+            }
+        }
+    }
+}
