+++ sample/src/main/java/com/mushafimad/sampleapp/component/DemoListItem.kt
@@ -1,0 +1,38 @@
+package com.mushafimad.sampleapp.component
+
+import androidx.compose.foundation.clickable
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ArrowForward
+import androidx.compose.material3.Icon
+import androidx.compose.material3.ListItem
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import com.mushafimad.sampleapp.modle.DemoItem
+
+@Composable
+fun DemoListItem(
+    item: DemoItem,
+    onClick: () -> Unit
+) {
+    ListItem(
+        headlineContent = { Text(item.title) },
+        supportingContent = { Text(item.subtitle) },
+        leadingContent = {
+            Icon(
+                imageVector = item.icon,
+                contentDescription = null,
+                tint = MaterialTheme.colorScheme.primary
+            )
+        },
+        trailingContent = {
+            Icon(
+                imageVector = Icons.Default.ArrowForward,
+                contentDescription = "Navigate",
+                tint = MaterialTheme.colorScheme.onSurfaceVariant
+            )
+        },
+        modifier = Modifier.Companion.clickable(onClick = onClick)
+    )
+}
