+++ sample/src/main/java/com/mushafimad/sampleapp/component/PreferenceItem.kt
@@ -1,0 +1,63 @@
+package com.mushafimad.sampleapp.component
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.KeyboardArrowDown
+import androidx.compose.material.icons.filled.KeyboardArrowUp
+import androidx.compose.material3.Card
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun PreferenceItem(
+    title: String,
+    value: String,
+    onIncrement: (() -> Unit)?,
+    onDecrement: (() -> Unit)?
+) {
+    Card(
+        modifier = Modifier.Companion
+            .fillMaxWidth()
+            .padding(vertical = 8.dp)
+    ) {
+        Row(
+            modifier = Modifier.Companion
+                .fillMaxWidth()
+                .padding(16.dp),
+            horizontalArrangement = Arrangement.SpaceBetween,
+            verticalAlignment = Alignment.Companion.CenterVertically
+        ) {
+            Column(modifier = Modifier.Companion.weight(1f)) {
+                Text(
+                    text = title,
+                    style = MaterialTheme.typography.labelMedium,
+                    color = MaterialTheme.colorScheme.primary
+                )
+                Text(
+                    text = value,
+                    style = MaterialTheme.typography.titleMedium
+                )
+            }
+            if (onDecrement != null && onIncrement != null) {
+                Row {
+                    IconButton(onClick = onDecrement) {
+                        Icon(Icons.Default.KeyboardArrowDown, "Decrease")
+                    }
+                    IconButton(onClick = onIncrement) {
+                        Icon(Icons.Default.KeyboardArrowUp, "Increase")
+                    }
+                }
+            }
+        }
+    }
+}
