+++ sample/src/main/java/com/mushafimad/sampleapp/component/ReciterItem.kt
@@ -1,0 +1,68 @@
+package com.mushafimad.sampleapp.component
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.unit.dp
+import com.mushafimad.core.domain.models.ReciterInfo
+
+@Composable
+fun ReciterItem(index: Int, reciter: ReciterInfo) {
+    Card(
+        modifier = Modifier.Companion
+            .fillMaxWidth()
+            .padding(vertical = 4.dp)
+    ) {
+        Row(
+            modifier = Modifier.Companion
+                .fillMaxWidth()
+                .padding(12.dp),
+            verticalAlignment = Alignment.Companion.CenterVertically
+        ) {
+            Box(
+                modifier = Modifier.Companion
+                    .size(40.dp)
+                    .clip(RoundedCornerShape(20.dp))
+                    .background(MaterialTheme.colorScheme.secondaryContainer),
+                contentAlignment = Alignment.Companion.Center
+            ) {
+                Text(
+                    text = "$index",
+                    style = MaterialTheme.typography.titleMedium,
+                    color = MaterialTheme.colorScheme.onSecondaryContainer
+                )
+            }
+            Spacer(modifier = Modifier.Companion.width(12.dp))
+            Column(modifier = Modifier.Companion.weight(1f)) {
+                Text(
+                    text = reciter.nameArabic,
+                    style = MaterialTheme.typography.titleMedium
+                )
+                Text(
+                    text = reciter.nameEnglish,
+                    style = MaterialTheme.typography.bodySmall,
+                    color = MaterialTheme.colorScheme.onSurfaceVariant
+                )
+                Text(
+                    text = "ID: ${reciter.id} | ${reciter.rewaya}",
+                    style = MaterialTheme.typography.labelSmall,
+                    color = MaterialTheme.colorScheme.tertiary
+                )
+            }
+        }
+    }
+}
