+++ sample/src/main/java/com/mushafimad/sampleapp/component/VerseItem.kt
@@ -1,0 +1,55 @@
+package com.mushafimad.sampleapp.component
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Card
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import com.mushafimad.core.domain.models.Verse
+
+@Composable
+ fun VerseItem(verse: Verse) {
+    Card(
+        modifier = Modifier.Companion
+            .fillMaxWidth()
+            .padding(vertical = 4.dp)
+    ) {
+        Column(
+            modifier = Modifier.Companion
+                .fillMaxWidth()
+                .padding(12.dp)
+        ) {
+            Row(
+                modifier = Modifier.Companion.fillMaxWidth(),
+                horizontalArrangement = Arrangement.SpaceBetween
+            ) {
+                Text(
+                    text = "${verse.chapterNumber}:${verse.number}",
+                    style = MaterialTheme.typography.labelMedium,
+                    color = MaterialTheme.colorScheme.primary
+                )
+                Text(
+                    text = "Page ${verse.pageNumber}",
+                    style = MaterialTheme.typography.labelSmall,
+                    color = MaterialTheme.colorScheme.onSurfaceVariant
+                )
+            }
+            Spacer(modifier = Modifier.Companion.height(8.dp))
+            Text(
+                text = verse.text,
+                style = MaterialTheme.typography.bodyLarge,
+                textAlign = TextAlign.Companion.End,
+                modifier = Modifier.Companion.fillMaxWidth()
+            )
+        }
+    }
+}
