+++ sample/src/main/java/com/mushafimad/sampleapp/modle/DemoItem.kt
@@ -1,0 +1,10 @@
+package com.mushafimad.sampleapp.modle
+
+import androidx.compose.ui.graphics.vector.ImageVector
+
+data class DemoItem(
+    val title: String,
+    val subtitle: String,
+    val icon: ImageVector,
+    val route: String
+)
