+++ sample/src/main/java/com/mushafimad/sampleapp/screens/MushafWithAudioScreen.kt
@@ -1,0 +1,50 @@
+package com.mushafimad.sampleapp.screens
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.navigation.NavHostController
+import com.mushafimad.core.domain.models.MushafType
+import com.mushafimad.ui.mushaf.MushafWithPlayerView
+import com.mushafimad.ui.theme.ColorSchemeType
+import com.mushafimad.ui.theme.ReadingTheme
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun MushafWithAudioScreen(navController: NavHostController) {
+    Scaffold(
+        topBar = {
+            TopAppBar(
+                title = { Text("MushafWithPlayerView") },
+                navigationIcon = {
+                    IconButton(onClick = { navController.popBackStack() }) {
+                        Icon(Icons.AutoMirrored.Filled.ArrowBack, "Back")
+                    }
+                }
+            )
+        }
+    ) { paddingValues ->
+        Box(modifier = Modifier.Companion.padding(paddingValues)) {
+            MushafWithPlayerView(
+                readingTheme = ReadingTheme.COMFORTABLE,
+                colorScheme = ColorSchemeType.DEFAULT,
+                mushafType = MushafType.HAFS_1441,
+                initialPage = 1,
+                showNavigationControls = true,
+                showPageInfo = true,
+                showAudioPlayer = true,
+                modifier = Modifier.Companion.fillMaxSize()
+            )
+        }
+    }
+}
