+++ sample/src/main/java/com/mushafimad/sampleapp/screens/ThemeCustomizationScreen.kt
@@ -1,0 +1,79 @@
+package com.mushafimad.sampleapp.screens
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material.icons.filled.Settings
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.ModalBottomSheet
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.navigation.NavHostController
+import com.mushafimad.core.domain.models.MushafType
+import com.mushafimad.sampleapp.component.ThemeSettingsSheet
+import com.mushafimad.ui.mushaf.MushafView
+import com.mushafimad.ui.theme.ColorSchemeType
+import com.mushafimad.ui.theme.ReadingTheme
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun ThemeCustomizationScreen(navController: NavHostController) {
+    var selectedTheme by remember { mutableStateOf(ReadingTheme.COMFORTABLE) }
+    var selectedColorScheme by remember { mutableStateOf(ColorSchemeType.DEFAULT) }
+    var showSettings by remember { mutableStateOf(false) }
+
+    Scaffold(
+        topBar = {
+            TopAppBar(
+                title = { Text("Theme System") },
+                navigationIcon = {
+                    IconButton(onClick = { navController.popBackStack() }) {
+                        Icon(Icons.AutoMirrored.Filled.ArrowBack, "Back")
+                    }
+                },
+                actions = {
+                    IconButton(onClick = { showSettings = true }) {
+                        Icon(Icons.Default.Settings, "Settings")
+                    }
+                }
+            )
+        }
+    ) { paddingValues ->
+        Box(modifier = Modifier.Companion.padding(paddingValues)) {
+            MushafView(
+                readingTheme = selectedTheme,
+                colorScheme = selectedColorScheme,
+                mushafType = MushafType.HAFS_1441,
+                initialPage = 1,
+                showNavigationControls = true,
+                showPageInfo = true,
+                modifier = Modifier.Companion.fillMaxSize()
+            )
+        }
+
+        if (showSettings) {
+            ModalBottomSheet(
+                onDismissRequest = { showSettings = false }
+            ) {
+                ThemeSettingsSheet(
+                    selectedReadingTheme = selectedTheme,
+                    selectedColorScheme = selectedColorScheme,
+                    onReadingThemeChange = { selectedTheme = it },
+                    onColorSchemeChange = { selectedColorScheme = it },
+                    onDismiss = { showSettings = false }
+                )
+            }
+        }
+    }
+}
